- some stuff
git-svn-id: http://svn.codergrid.de/svn/sf/bruteripper/trunk@346 7add9c9f-332d-0410-adcc-f7b1da7091d4

+++ src/org/esb/hive/job/ClientHandler.cpp
@@ -87,8 +87,8 @@ bool ClientHandler::putProcessUnit(ProcessUnit & unit){
 		int  field=0;
 
         _stmt->setInt( "stream_id", unit._target_stream);
-        _stmt->setInt( "pts", packet->packet->pts);
-        _stmt->setInt( "dts", packet->packet->dts);
+        _stmt->setDouble( "pts", packet->packet->pts);
+        _stmt->setDouble( "dts", packet->packet->dts);
         _stmt->setInt( "stream_index", packet->packet->stream_index);
         _stmt->setInt( "key_frame", packet->isKeyFrame());
 		if(packet->packet->stream_index==0)
@@ -97,7 +97,7 @@ bool ClientHandler::putProcessUnit(ProcessUnit & unit){
     	    _stmt->setInt( "frame_group",0);	
         _stmt->setInt( "flags", packet->packet->flags);
         _stmt->setInt( "duration", packet->packet->duration);
-        _stmt->setInt( "pos", packet->packet->pos);
+        _stmt->setDouble( "pos", packet->packet->pos);
         _stmt->setInt( "data_size", packet->packet->size);
         _stmt->setBlob( "data", (char *)packet->packet->data,packet->packet->size);
 		_stmt->execute();
